projects/01_adaptive_rag/requirements.txt
langgraph>=0.2
langchain-openai>=0.2
langchain-community>=0.3
langchain-text-splitters>=0.3
chromadb>=0.5
tavily-python>=0.5
python-dotenv>=1.0
